requirements.txt
pywin32==310
toml==0.10.2
gitpython==3.1.31 